requirements.txt
pygame-ce==2.5.1
numpy==2.1.0
